requirements.txt
google-adk>=1.28.0
google-cloud-aiplatform>=1.132.0
google-cloud-alloydb-connector[pg8000]>=1.4.0
google-cloud-firestore==2.19.0
google-cloud-webrisk>=1.14.0
google-cloud-language>=2.15.0
google-cloud-translate>=3.15.0
google-cloud-storage>=2.18.0
google-cloud-logging>=3.11.0
google-cloud-secret-manager>=2.21.0
google-auth>=2.38.0
google-auth-oauthlib>=1.2.2
google-api-python-client>=2.170.0
google-auth-httplib2>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
androguard>=4.1.2
requests>=2.32.0
python-dotenv>=1.0.1
python-whois==0.9.5
dnspython==2.6.1
pg8000>=1.31.2
aiohttp>=3.11.0
httpx>=0.28.0
pydantic>=2.10.0
Pillow>=11.1.0
google-genai==1.70.0
websockets>=15.0.1,<16.0.0
